Refactor CommandExecutor: extract protocol and fix SwiftLint violations
- Extract CommandExecuting protocol to separate file for better organization
- Fix test failure by filtering shell-init errors in CommandExecutorTests
- Fix all 28 SwiftLint violations (trailing whitespace, function length, etc)
- Refactor large executeProcessWithStreamingAsync function into smaller helpers
- Merge back processStreamsWithTaskGroup to reduce parameter count
- Maintain all functionality while improving code quality
