+++ Sources/Xproject/Utilities/CommandExecutor.swift
@@ -5,46 +5,6 @@
 
 import Foundation
 
-// MARK: - Command Execution Protocol
-
-public protocol CommandExecuting: Sendable {
-    func execute(_ command: String, workingDirectory: URL?, environment: [String: String]?) throws -> CommandResult
-    func executeOrThrow(_ command: String, workingDirectory: URL?, environment: [String: String]?) throws -> CommandResult
-    func executeReadOnly(_ command: String, workingDirectory: URL?, environment: [String: String]?) throws -> CommandResult
-    func executeWithStreamingOutput(_ command: String, workingDirectory: URL?, environment: [String: String]?) async throws -> CommandResult
-    func commandExists(_ command: String) -> Bool
-}
-
-public extension CommandExecuting {
-    func execute(_ command: String, workingDirectory: URL? = nil, environment: [String: String]? = nil) throws -> CommandResult {
-        return try execute(command, workingDirectory: workingDirectory, environment: environment)
-    }
-
-    func executeOrThrow(_ command: String, workingDirectory: URL? = nil, environment: [String: String]? = nil) throws -> CommandResult {
-        let result = try execute(command, workingDirectory: workingDirectory, environment: environment)
-
-        if result.exitCode != 0 {
-            throw CommandError.executionFailed(result: result)
-        }
-
-        return result
-    }
-
-    func executeReadOnly(_ command: String, workingDirectory: URL? = nil, environment: [String: String]? = nil) throws -> CommandResult {
-        return try execute(command, workingDirectory: workingDirectory, environment: environment)
-    }
-
-    func executeWithStreamingOutput(
-        _ command: String,
-        workingDirectory: URL? = nil,
-        environment: [String: String]? = nil
-    ) async throws -> CommandResult {
-        return try await executeWithStreamingOutput(command, workingDirectory: workingDirectory, environment: environment)
-    }
-}
-
-// MARK: - Command Execution Utilities
-
 public struct CommandExecutor: CommandExecuting, Sendable {
     private let dryRun: Bool
 
@@ -190,7 +150,6 @@ public struct CommandExecutor: CommandExecuting, Sendable {
         )
     }
 
-
     /// Create and configure a Process for command execution
     private func createProcess(
         command: String,
@@ -245,7 +204,6 @@ public struct CommandExecutor: CommandExecuting, Sendable {
         )
     }
 
-
     /// Create CommandResult from process and data
     private func createCommandResult(
         from process: Process,
@@ -280,7 +238,7 @@ public struct CommandExecutor: CommandExecuting, Sendable {
         if let workingDirectory = workingDirectory {
             context += " (in \(workingDirectory.path))"
         }
-        
+
         print("$ \(command)\(context)")
 
         let process = createProcess(command: command, workingDirectory: workingDirectory, environment: environment)
@@ -293,68 +251,15 @@ public struct CommandExecutor: CommandExecuting, Sendable {
             command: command
         )
     }
-    
+
     /// Execute process with streaming output using AsyncStream
     private func executeProcessWithStreamingAsync(_ process: Process) async -> (Data, Data) {
-        let outputPipe = Pipe()
-        let errorPipe = Pipe()
-        process.standardOutput = outputPipe
-        process.standardError = errorPipe
-        
-        // Use actor for thread-safe data collection
-        actor DataCollector {
-            var outputData = Data()
-            var errorData = Data()
-            
-            func appendOutput(_ data: Data) {
-                outputData.append(data)
-            }
-            
-            func appendError(_ data: Data) {
-                errorData.append(data)
-            }
-            
-            func getData() -> (Data, Data) {
-                (outputData, errorData)
-            }
-        }
-        
+        let (outputPipe, errorPipe) = setupProcessPipes(process)
         let collector = DataCollector()
-        
-        // Create async streams for output and error
-        let outputStream = AsyncStream<Data> { continuation in
-            outputPipe.fileHandleForReading.readabilityHandler = { handle in
-                let data = handle.availableData
-                if data.isEmpty {
-                    continuation.finish()
-                } else {
-                    continuation.yield(data)
-                }
-            }
-            
-            continuation.onTermination = { _ in
-                outputPipe.fileHandleForReading.readabilityHandler = nil
-            }
-        }
-        
-        let errorStream = AsyncStream<Data> { continuation in
-            errorPipe.fileHandleForReading.readabilityHandler = { handle in
-                let data = handle.availableData
-                if data.isEmpty {
-                    continuation.finish()
-                } else {
-                    continuation.yield(data)
-                }
-            }
-            
-            continuation.onTermination = { _ in
-                errorPipe.fileHandleForReading.readabilityHandler = nil
-            }
-        }
-        
-        // Start the process
+        let (outputStream, errorStream) = createAsyncStreams(outputPipe: outputPipe, errorPipe: errorPipe)
+
         try? process.run()
-        
+
         // Process streams concurrently
         await withTaskGroup(of: Void.self) { group in
             group.addTask {
@@ -366,7 +271,7 @@ public struct CommandExecutor: CommandExecuting, Sendable {
                     await collector.appendOutput(data)
                 }
             }
-            
+
             group.addTask {
                 for await data in errorStream {
                     if let string = String(data: data, encoding: .utf8) {
@@ -376,7 +281,7 @@ public struct CommandExecutor: CommandExecuting, Sendable {
                     await collector.appendError(data)
                 }
             }
-            
+
             group.addTask {
                 process.waitUntilExit()
                 // Ensure streams are closed when process exits
@@ -384,15 +289,63 @@ public struct CommandExecutor: CommandExecuting, Sendable {
                 errorPipe.fileHandleForReading.readabilityHandler = nil
             }
         }
-        
-        // Collect any remaining data
+
+        await collectRemainingData(outputPipe: outputPipe, errorPipe: errorPipe, collector: collector)
+
+        return await collector.getData()
+    }
+
+    /// Setup pipes for process output and error streams
+    private func setupProcessPipes(_ process: Process) -> (Pipe, Pipe) {
+        let outputPipe = Pipe()
+        let errorPipe = Pipe()
+        process.standardOutput = outputPipe
+        process.standardError = errorPipe
+        return (outputPipe, errorPipe)
+    }
+
+    /// Create async streams for output and error data
+    private func createAsyncStreams(outputPipe: Pipe, errorPipe: Pipe) -> (AsyncStream<Data>, AsyncStream<Data>) {
+        let outputStream = AsyncStream<Data> { continuation in
+            outputPipe.fileHandleForReading.readabilityHandler = { handle in
+                let data = handle.availableData
+                if data.isEmpty {
+                    continuation.finish()
+                } else {
+                    continuation.yield(data)
+                }
+            }
+
+            continuation.onTermination = { _ in
+                outputPipe.fileHandleForReading.readabilityHandler = nil
+            }
+        }
+
+        let errorStream = AsyncStream<Data> { continuation in
+            errorPipe.fileHandleForReading.readabilityHandler = { handle in
+                let data = handle.availableData
+                if data.isEmpty {
+                    continuation.finish()
+                } else {
+                    continuation.yield(data)
+                }
+            }
+
+            continuation.onTermination = { _ in
+                errorPipe.fileHandleForReading.readabilityHandler = nil
+            }
+        }
+
+        return (outputStream, errorStream)
+    }
+
+    /// Collect any remaining data from pipes after process completion
+    private func collectRemainingData(outputPipe: Pipe, errorPipe: Pipe, collector: DataCollector) async {
         let remainingOutputData = outputPipe.fileHandleForReading.readDataToEndOfFile()
         let remainingErrorData = errorPipe.fileHandleForReading.readDataToEndOfFile()
-        
+
         await collector.appendOutput(remainingOutputData)
         await collector.appendError(remainingErrorData)
-        
-        return await collector.getData()
     }
 
     /// Check if a command exists in PATH
@@ -448,5 +401,25 @@ public enum CommandError: Error, LocalizedError, Sendable {
         case .commandNotFound(let command):
             return "Command '\(command)' not found in PATH"
         }
+    }
+}
+
+// MARK: - Command Execution Utilities
+
+/// Actor for thread-safe data collection during streaming command execution
+private actor DataCollector {
+    var outputData = Data()
+    var errorData = Data()
+
+    func appendOutput(_ data: Data) {
+        outputData.append(data)
+    }
+
+    func appendError(_ data: Data) {
+        errorData.append(data)
+    }
+
+    func getData() -> (Data, Data) {
+        (outputData, errorData)
     }
 }
